evolv_backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.1.6
django-cors-headers==4.9.0
django-countries==7.6.1
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.2.0
